Conceal no-break spaces around emphasized text
This was my only item on my wishlist from the very begining but
the point of slides.vim is to not spend time on it.

But today, @habamax (also habamax libera #vim channel) made it
so i can write this very grateful commit message.

The bad news is: you need to fix all your slides. this command made the
job for me:

    locate -0b '*.slides'|
        xargs -n0 sed -i.oldemphasize -r 's/ [^ ]+ / & /g'